Renamed implementation of service, split interface from implementation in two files. Split out a method.
